0.2.0: Added more intrinsic functions and the output definition.
Added the majority of the intrinsic functions, maps and related functions
have been excluded fornow as I can't really think of a use for them. We'll
add them later if we need them.

Documentation has been updated to reflect these changes. I've added the concept
of a 'definition', if I like this later on we should probably make a few code
changes for consistency.